requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
markdownify>=0.11.6
html2text>=2020.1.16
selenium>=4.15.0
webdriver-manager>=4.0.0